`Reorganize HomePage file structure and update imports`
The HomePage component was moved to a `pages` directory, requiring updates to imports across relevant files. Adjusted asset and dependency paths to reflect the new structure. These changes improve project organization and maintain consistency.
