+++ cypress/component/HomePage.cy.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { mount } from '@cypress/react';
-import HomePage from '../../src/HomePage';
+import HomePage from '../../src/pages/HomePage';
 
 describe('HomePage Component Tests', () => {
   beforeEach(() => {
