+++ src/pages/HomePage.tsx
@@ -4,19 +4,19 @@ import { Link } from 'react-router-dom';
 import { Trans, useTranslation } from 'react-i18next';
 
 // Material UI imports
-import { Avatar, Box, Button, Paper, Typography, alpha, styled, useMediaQuery, useTheme } from '@mui/material';
+import { alpha, Avatar, Box, Button, Paper, styled, Typography, useMediaQuery, useTheme } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
 import Grid from '@mui/material/Grid2';
 
 // Local data
-import BackgroundImage from './images/background.jpg';
-import GreenSwirl1Image from './images/green-swirl1.png';
-import GreenSwirl2Image from './images/green-swirl2.png';
-import Decal1Image from './images/decal1.svg';
-import Decal2Image from './images/decal2.svg';
-import Decal2DarkImage from './images/decal2-dark.svg';
-import Decal2DarkHCImage from './images/decal2-darkhc.svg';
-import FacilityImage from './images/facility.jpg';
+import BackgroundImage from '../images/background.jpg';
+import GreenSwirl1Image from '../images/green-swirl1.png';
+import GreenSwirl2Image from '../images/green-swirl2.png';
+import Decal1Image from '../images/decal1.svg';
+import Decal2Image from '../images/decal2.svg';
+import Decal2DarkImage from '../images/decal2-dark.svg';
+import Decal2DarkHCImage from '../images/decal2-darkhc.svg';
+import FacilityImage from '../images/facility.jpg';
 
 const backgroundTitleStyles = {
   color: '#FFFFFF',
